Address full-pass findings: kind warning, numeric pct_used, hypopg default, doc/script hygiene
- schemas.ts: pg_describe_table emits a warning when the kind fetch fails
  instead of silently reporting "table" for what may be a view.
- admin.ts: pg_advisor sequence_exhaustion divides in numeric, not float8;
  preserves precision past 2^53 where the >= threshold check actually matters.
- explain.ts: buildHypopgHooks defaults `using` to "btree" so handler-direct
  calls that bypass Zod don't render `USING undefined`. Integration regression
  test added for the omitted-using path.
- api.ts: documents getPool() env-var snapshot semantics and typeNameCache
  staleness bound so the hidden assumptions are visible.
- wsl-pg-setup.sh: DEV/CI ONLY warnings on the PG16 purge and the
  0.0.0.0/0 md5 pg_hba stanza -- safe inside WSL, lethal as a template.
- release.sh: Verify step uses the same 5x5s retry loop as the CI smoke test
  instead of a one-shot sleep that flakes on a slow registry.

+++ src/tools/admin.ts
@@ -357,16 +357,19 @@ export const adminTools = [
           }>(
             // pg_sequences was added in PG10. last_value can be NULL on a never-
             // touched sequence; we filter those out (nothing to report yet).
+            // Divide in `numeric`, not `float8`: BIGINT sequences past 2^53 lose
+            // precision in float8, and the danger zone (>= threshold) is exactly
+            // where the reported pct_used must stay accurate.
             `SELECT
                schemaname AS schema,
                sequencename AS sequence,
                last_value::text AS last_value,
                max_value::text AS max_value,
-               (last_value::float8 / NULLIF(max_value::float8, 0))::numeric(6, 4)::float8 AS pct_used
+               (last_value::numeric / NULLIF(max_value::numeric, 0))::numeric(6, 4)::float8 AS pct_used
              FROM pg_catalog.pg_sequences
              WHERE last_value IS NOT NULL
                AND max_value > 0
-               AND (last_value::float8 / max_value::float8) >= $1
+               AND (last_value::numeric / max_value::numeric) >= $1
              ORDER BY pct_used DESC NULLS LAST
              LIMIT $2`,
             [seqExhaustionThreshold, limit],
